+++ CHANGELOG.md
@@ -18,6 +18,13 @@ Types of changes:
 
 ### Added
 
+## [0.1.04] - 2025-09-01
+
+### Added
+
+- Now instance scan is compatible with the new public folder structure.
+- A new style for issues in integration.
+
 ## [0.1.03] - 2024-12-02
 
 ### Added